streaming/requirements.txt
polygon-api-client>=1.14.0
snowpipe-streaming>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
